Extract room request button to separate component
